+++ CHANGES.md
@@ -21,6 +21,8 @@
 -   Added ability to add fully customised affinity to statefulsets in helm
 -   Fixed docker image build script failed on windows platform
 -   Only runtime dependencies will be included by docker image build script
+-   Added `cloudsql-db-credentials` to create-secrets tool
+-   Fixed an file selector error when current directory & non of its sub directory has \*.json file
 
 ## 0.0.50